UBUNTU: custom binaries need VERSION_SIGNATURE updated during prepare
BugLink: http://bugs.launchpad.net/bugs/794698

When preparing the custom binary trees we build a configuration.  Part
of this should involve setting VERSION_SIGNATURE for that build.  This
is missed for custom binaries which leads to the confusing situation
that a -xen flavour will report as below:

    $ uname -a && cat /proc/version_signature
    Linux ip-10-36-7-174 2.6.24-29-xen #1 SMP
			Thu Apr 21 20:44:30 UTC 2011 i686 GNU/Linux
    Ubuntu 2.6.24-4.6-generic

This renders /proc/version_signature worthless.  Ensure we fill this
in when we build the config.

Acked-by: Tim Gardner <tim.gardner@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>