UBUNTU: (config) Disable COMPAT_VDSO for i386 Xen kernels
BugLink: http://bugs.launchpad.net/bugs/794715

This option is already disabled for generic i386 kernels as it
practically renders randomization of the VDSO location useless.
According to the description of this option it only is needed for
older (<2.3.3) versions of libc. As the Hardy version is newer
than that, it makes no sense to keep it enabled.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Andy Whitcroft <andy.whitcroft@canonical.com>
Acked-by: Tim Gardner <tim.gardner@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>