sru-workflow-manager: Added bug finalization after publishing
When the tasks for publishing to -updates and optionally to -security
have been set to "Fix Released", then we send email and a status update,
and set the master workflow task to "Fix Released". This closes it, and
also prevents the workflow bot from processing it again.

I also consolidated some of the status update code and made the text
in the updates always include the package name and version

Signed-off-by: Steve Conklin <sconklin@canonical.com>