ktl: simplify and fix debian changelog detection
The debian changelog is in one of three places.  It is either in a branch
specific changelog, debian/changelog, or meta-source/debian/changelog.
Where a branch specific changelog is present it is always identified via
the debian/debian.env DEBIAN=<directory> stanza.  Move to using this to
identify the correct debian changelog.  Also move to using a loop over
the possible locations.

Signed-off-by: Andy Whitcroft <apw@canonical.com>