Add ability to customize task assignments based on package name
Some packages require assignment to different teams for some
tasks. Task assignment lists are now create per package name,
with a default list for packages not specifically configured.

Also, a tag was added for ARM packages. This perhaps should
be abstracted in the same way that the assigmnents were, but
this was a simple solution.

Signed-off-by: Steve Conklin <sconklin@canonical.com>