UBUNTU: commonise the rebase scripting
Move to the new common rebase scripting framework.  This does not make
a new remote when pulling in the updates.  This avoids corrupting a
local remote if it already exists, limits the fetch only to the required
branch, and finally stops this repo being polluted by unnecessary tags.
Significantly improves rebase performance as a result.

Signed-off-by: Andy Whitcroft <apw@canonical.com>