bugs: proc-cve-bugs -- invalidate non-supported conbinations
CVEs are created against numerous source packages and nominated for all
series.  For a lot of them we know that the combinations are invalid.
For example, the LTS backports kernels are only valid on Lucid.  Use the
information encoded in the ktl/ubuntu object to determine which are
invalid or unsupported and close them out agressivly.

Signed-off-by: Andy Whitcroft <apw@canonical.com>