close races in /proc/*/{environ,auxv}, CVE-2011-1020
Equivalent to:
	commit ec6fd8a4355cda81cd9f06bebc048e83eb514ac7
	commit d6f64b89d7ff22ce05896ab4a93a653e8d0b123d
	commit 2fadaef41283aad7100fa73f01998cddaca25833

Note that the locking is looser than that in upstream this may leaves
a small window which may be exploitable, that said the locks which are
used to ensure there is no window are not in existance back in hardy.
Indeed the races in exec which they are introduced to fix are also present.

CVE-2011-1020
BugLink: http://bugs.launchpad.net/bugs/813026
Signed-off-by: Andy Whitcroft <apw@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>