stable/sru-workflow-manager: initialize taskbyname for each processed bug
We may now have different sets of tasks for each bug (like an extra
upload-to-ppa or prepare-package-* task). So the taskbyname dictionary
will vary in size, and if we don't initialize for each bug, we may end
up changing tasks from previous processed bugs, as code now must make
assumptions based on tasks present in a bug.

Signed-off-by: Herton Ronaldo Krzesinski <herton.krzesinski@canonical.com>