Add support to verification-reverted-* tags
Currently, it's difficult to determine when a bug got its SRU patch
reverted when looking at sru-report page: after verification failed, the
failed status remains until that kernel is released. With this change,
we introduce support to a new verification-reverted-<series> tag, which
will be added to bug reports that got its patch reverted because of a
previous verification failure, so it's easy to process and see them on
sru-report page.

Signed-off-by: Herton Ronaldo Krzesinski <herton.krzesinski@canonical.com>