Added new status text for test failures and process holds
Added new text of TestFail for status when cert or qa testing
is complete and failed. In those cases we also stop processing
the tracking bug by setting the master task to invalid.

Also added new text of OnHold for when we run into process problems
and stop processing.

For the cases when we detect that cert or qa tasks have been set to
complete but the proper tags have not been applied, we no longer
stop processing of the bug, we simply reset the status for that
cert or qa task to Incomplete and make a comment entry, so that
the responsible team can fix it.