SGX: Enable early suspend flag for all OMAP3/4 builds
This patch enables the SGX_EARLYSUSPEND flag on all the OMAP3/4
builds which allows the SGX display driver to receive properly
the suspend/resume requests from the early suspend mechanism.
This will solve suspend/resume 3D application issues
when the dual display feature is enabled in the DDK.
Currently it was only enabled on OMAP3 builds.

Signed-off-by: Gustavo Diaz Prado <x0083741@ti.com>
Change-Id: Ief6e221f916729a0747e81b550f6fb426184a61b