drv/i915: Pull display_clock_mode out of VBT table
BugLink: http://bugs.launchpad.net/bugs/614238

This tells the driver whether a CK505 clock source is available on
pre-PCH hardware. If so, it should be used as the non-SSC source,
leaving the internal clock for use as the SSC source.

Signed-off-by: Keith Packard <keithp@keithp.com>
Reviewed-by: Chris Wison <chris@chris-wilson.co.uk>
(cherry picked from commit abd06860185fc613776adae792772e076d77caf2)

Signed-off-by: Tim Gardner <tim.gardner@canonical.com>
Acked-by: Herton Krzesinski <herton.krzesinski@canonical.com>
Acked-by: Seth Forshee <seth.forshee@canonical.com>