ktl/ubuntu.py: add series_name function
The series_name function returns the name of the series, based on the
package and version passed to it. It isn't a simply version -> series
match, as we have several especial cases, for example, 2.6.31 isn't
always karmic, it can be linux-fsl-imx51 for lucid. The same can be said
for linux-mvl-dove, where we have 2.6.32 versions on maverick, not only
on lucid.

It was copied and adapted from expected_series_name at workflow.py, and
is useful for many tools which needs to get or check the series based on
a package-version match.

Signed-off-by: Herton Ronaldo Krzesinski <herton.krzesinski@canonical.com>