UBUNTU: [Config] Use debian.env earlier
Until now, we were creating debian.env upon first debian/rules run. This
is incompatible with our tools, which expects debian.env to be available
always, commited in the repository. This is a minimal change to have
always debian.env present (ideally we could move debian.fsl-imx51/rules
to debian/rules, but not doing this at this stage for fsl-imx51).

Acked-by: Tim Gardner <tim.gardner@canonical.com>
Acked-by: Andy Whitcroft <apw@canonical.com>
Signed-off-by: Herton Ronaldo Krzesinski <herton.krzesinski@canonical.com>