stable/sru-workflow-manager: don't crash when opening tracking bugs
Avoid crashing when opening tracking bugs, otherwise on next run we will
process the same state as before (which may again open a new tracking
bug for something already processed).

The solution here is just catching the exception when trying to open a
new tracking bug. If creating the bug failed, comment on main bug that
it wasn't possible to open a new tracking bug for a derivative or
backport package, so it can be manually done later.

We should protect any state/bug changes to avoid the bug or
sru-workflow-manager be in an inconsistent state, but for now just
creating the bug is protected, as we have an expected point of failure
with it: launchpad doesn't allow a bug to be created for a package when
it was never published, and gives an error if this is tried (for which
an exception is raised). We should catch this, for example, with any
linux-lts-backport-<latest> at the start of a new cycle.

This isn't perfect, but there is no perfect solution, since Launchpad
ideally should support commit/rollback of changes (in this way, for each
state sru-workflow-manager process, it would queue up and commit all
changes in the end, avoiding staying in a inconsistent state, and
rolling back for any error it gets).

Signed-off-by: Herton Ronaldo Krzesinski <herton.krzesinski@canonical.com>