buildscripts/target-scripts/run-build: Background dpkg-buildpackage
This is required because the code later on waits for that job
to complete. If the main build command does not run in the background,
then $! is empty and unfortunatly wait without argument waits for all
jobs sent into the background, which is then a "tail -f" and so the
build script never stops.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>