devel/reconcile-generic: adjust counters for ubuntu, debian, and sauce
reconcile-generic generates a summary similar to:

=== total: 341 ignored: 32 (upstreamed: 0 reverted: 32) needed: 309 (debian: 127 ubuntu: 76 sauce: 122) ===

We see that total-ignored=needed

needed is made up of "debian", "ubuntu", and "sauce" patches.  However,
according to the above numbers, we see that debian+ubuntu+sauce!=needed.
This is because the "debian", "ubuntu", and "sauce" patches were still
accounting for ignored patches.

This patch fixes up the accounting to not include ignored patches so
that debian+ubuntu+sauce=needed.  ie,

=== total: 341 ignored: 32 (upstreamed: 0 reverted: 32) needed: 309 (debian: 126 ubuntu: 65 sauce: 118) ===

Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>